# Conflicts:
#	pom.xml
#	src/main/kotlin/no/nav/bidrag/behandling/database/datamodell/Behandling.kt
#	src/main/kotlin/no/nav/bidrag/behandling/database/datamodell/Rolle.kt
#	src/main/kotlin/no/nav/bidrag/behandling/dto/v1/behandling/BehandlingDto.kt
#	src/main/kotlin/no/nav/bidrag/behandling/service/BeregningService.kt
#	src/main/kotlin/no/nav/bidrag/behandling/service/PersonService.kt
#	src/main/kotlin/no/nav/bidrag/behandling/service/VedtakService.kt
#	src/main/kotlin/no/nav/bidrag/behandling/transformers/BeregningDtoMapping.kt
#	src/main/kotlin/no/nav/bidrag/behandling/transformers/vedtak/mapping/fravedtak/VedtakTilBehandlingMapping.kt
#	src/main/kotlin/no/nav/bidrag/behandling/transformers/vedtak/mapping/tilvedtak/GrunnlagByggerBidrag.kt
#	src/test/kotlin/no/nav/bidrag/behandling/service/VedtakserviceBidragTest.kt

+++ src/main/kotlin/no/nav/bidrag/behandling/controller/v2/BehandlingControllerV2.kt
@@ -354,7 +354,7 @@ class BehandlingControllerV2(
         @Valid
         @RequestBody(required = true)
         opprettBehandling: OpprettBehandlingFraVedtakRequest,
-        @PathVariable refVedtaksId: Long,
+        @PathVariable refVedtaksId: Int,
     ): OpprettBehandlingResponse = vedtakService.opprettBehandlingFraVedtak(opprettBehandling, refVedtaksId)
 
     @Suppress("unused")
